Use uint64 instead of casting to uint
This fixes a code-scanning warning [1] for converting between integer
types without checking for the upper bound.

[1] https://github.com/SUSE/elemental/security/code-scanning/3

+++ internal/image/install/install.go
@@ -36,9 +36,9 @@ func (d DiskSize) IsValid() bool {
 }
 
 // returns the size in MiB
-func (d DiskSize) ToMiB() (uint, error) {
+func (d DiskSize) ToMiB() (uint64, error) {
 	size := strings.TrimSpace(string(d))
-	dimension := uint(1)
+	dimension := uint64(1)
 	switch unicode.ToUpper(rune(size[len(size)-1])) {
 	case 'K':
 		dimension = units.KiB
@@ -55,7 +55,7 @@ func (d DiskSize) ToMiB() (uint, error) {
 		return 0, err
 	}
 
-	return uint(value) * dimension / units.MiB, nil
+	return value * dimension / units.MiB, nil
 }
 
 type Installation struct {
